refactor(units): SI substance amount header (#4)
